GUI now attempts to fix mixed up soundpack directories (usually caused by a system crash) on startup.
Made DfoDir, CustomSoundpackDir, and TempSoundpackDir in the GUI main form go to and from member variables instead of directly to m_launcher. This makes them behave like the other properties and will make it easier to give UI for them if that is ever needed.

Created SetLauncherParams() in the main form so that code is only in one place.