First crack at the registry bug on 64-bit Windows. Dfo.Controlling.dll is set to compile with platform=x86 in all configurations. DFOCP.exe is compiled with the solution's platform. Let's see if DFOCP.exe with Any CPU and Dfo.Controlling.dll with x86 works. DFOCP.exe with x86 should work, but it would be nice if clients of Dfo.Controlling.dll could compile with Any CPU but always run in x86 mode.
DFO and its installer are 32-bit apps. When running on 64-bit Windows, 32-bit applications that look for registry keys in HKEY_LOCAL_MACHINE\SOFTWARE get transparently redirected to HKEY_LOCAL_MACHINE\SOFTWARE\Wow6432Node. DFOCP runs as a 64-bit application on 64-bit operating systems that have the 64-bit CLR installed so it does not get redirected. If DFOCP can be run as a 32-bit application, it should get the redirection which is what we want.

An alternative approach would be to check if we're running as a 64-bit app, and if we are, look for RootPath in HKEY_LOCAL_MACHINE\SOFTWARE\Wow6432Node\Nexon\DFO instead of HKEY_LOCAL_MACHINE\SOFTWARE\Nexon\DFO. This approach seems less portable than simply running as a 32-bit app.