Switchable files now declare whether they are a file or a directory so code knows whether to use File.Move or Directory.Move.
Removed some setters from ISwitchableFile that don't need to be there. In particular, it didn't make a lot of sense to have a setter for the Switch property in CheckboxSwitchableFile. What should it do if the checkbox is disabled because the files aren't switchable? It could enable the checkbox to force setting the value, but callers really shouldn't be messing with the enabled-ness of the checkbox. It could only change the checkbox if it is enabled, but then doing a set followed by a get might not get the same value you set.

Some documentation and stuff. Removed some dead code.

Fixed a bug in the command-line arguments for switchable files relating to variable capturing in lambda expressions.