Show in taskbar when restoring from tray so the "restore" window animation happens.
Automatically minimize to tray when the game starts actually launching. Restore if state becomes None and the app isn't exiting.

Hide tray icon when the app is exiting so it doesn't linger after the app is gone.

Fixed Dispose() in the main form because things in the form weren't actually getting disposed.

Updated TODO. Minimize to tray is done. Icon is "done", at least as well as I can do without the help of an artist.