Added support for clients of Dfo.Controlling to specify how much, if any, sensitive information like usernames and passwords they want to be logged.
Added logging to DfoLogin.cs. Logging is done "internally" rather than "externally" - that is, methods that do a significant amount of work log their beginning and end and anything important in the middle, rather than having callers logging the beginning and end of significant operations. This prevents disconnectedness in the log files from if callers don't log the beginning of an operation then a seemlingly unrelated log message from the middle of the called method shows up. It also means fewer log statements because the logging happens in one place instead of many. Exceptions still have to be logged by the catcher, not the thrower.

The DFOCP code will have to be changed because many places used a more "external" logging model.